refactor: split terrain tileset into library parts (#3878)
Extract Wang tile models and L2 variant key resolution from the 464-line
terrain_tileset.dart into dedicated part files for Phase 3 flame map
modularization; TerrainTilesetCache behavior unchanged.
