+++ app/lib/features/game/flame/tilesets/terrain_tileset.dart
@@ -11,22 +11,10 @@ import 'package:flutter/services.dart';
 
 import '../caches/asset_image_cache.dart';
 
+part 'terrain_tileset_models.dart';
+part 'terrain_tileset_variant_keys.dart';
+
 final _log = packageLogger();
-
-/// Sea terrain identifier (not in TerrainType enum).
-const String seaTerrainId = 'sea';
-
-/// Plains terrain identifier for L1 land base.
-const String plainsTerrainId = 'plains';
-
-/// Desert terrain identifier for L1 land base.
-const String desertTerrainId = 'desert';
-
-/// Terrain layer for the layered rendering architecture.
-/// L0: Sea (Wang tilesets for coastline).
-/// L1: Plains and Desert (Wang tilesets for land transitions).
-/// L2+: Features (standalone overlay tiles).
-enum TerrainLayer { layer0Sea, layer1LandBase, layer2Features }
 
 // Forest is split into hardwood and scrub variants (issue #3573 R8/R9). The
 // standalone tile keys mirror `tile_${TerrainType.name}` so
@@ -46,198 +34,6 @@ const String _tileSwamp = 'tile_swamp';
 const String _tilePlainsGrain = 'tile_plains_grain';
 const String _tilePlainsMeat = 'tile_plains_meat';
 const String _tilePlainsHorses = 'tile_plains_horses';
-
-/// Determines the rendering layer for a terrain type.
-/// Desert is L1 (land base alongside plains), not L2.
-TerrainLayer terrainLayer(TerrainType terrain) {
-  switch (terrain) {
-    case TerrainType.plains:
-    case TerrainType.desert:
-      return TerrainLayer.layer1LandBase;
-    case TerrainType.hardwoodForest:
-    case TerrainType.scrubForest:
-    case TerrainType.hills:
-    case TerrainType.mountain:
-    case TerrainType.swamp:
-      return TerrainLayer.layer2Features;
-  }
-}
-
-/// Tile metadata from PixelLab Wang tileset JSON.
-class WangTile {
-  final String id;
-  final Map<String, String> corners;
-  final Rect boundingBox;
-
-  WangTile({
-    required this.id,
-    required this.corners,
-    required this.boundingBox,
-  });
-
-  factory WangTile.fromJson(Map<String, dynamic> json) {
-    return WangTile(
-      id: json['id'] as String,
-      corners: Map<String, String>.from(
-        json['corners'] as Map<dynamic, dynamic>,
-      ),
-      boundingBox: Rect.fromLTWH(
-        (json['bounding_box']['x'] as num).toDouble(),
-        (json['bounding_box']['y'] as num).toDouble(),
-        (json['bounding_box']['width'] as num).toDouble(),
-        (json['bounding_box']['height'] as num).toDouble(),
-      ),
-    );
-  }
-}
-
-/// Loaded Wang tileset with image and tile metadata.
-class WangTileset {
-  final String name;
-  final String lowerTerrainId;
-  final String upperTerrainId;
-  final String? lowerBaseTileId;
-  final String? upperBaseTileId;
-  final ui.Image image;
-  final List<WangTile> tiles;
-
-  WangTileset({
-    required this.name,
-    required this.lowerTerrainId,
-    required this.upperTerrainId,
-    this.lowerBaseTileId,
-    this.upperBaseTileId,
-    required this.image,
-    required this.tiles,
-  });
-
-  WangTile? findTile({
-    required bool nw,
-    required bool ne,
-    required bool sw,
-    required bool se,
-  }) {
-    final nwCorner = nw ? 'upper' : 'lower';
-    final neCorner = ne ? 'upper' : 'lower';
-    final swCorner = sw ? 'upper' : 'lower';
-    final seCorner = se ? 'upper' : 'lower';
-
-    for (final tile in tiles) {
-      if (tile.corners['NW'] == nwCorner &&
-          tile.corners['NE'] == neCorner &&
-          tile.corners['SW'] == swCorner &&
-          tile.corners['SE'] == seCorner) {
-        return tile;
-      }
-    }
-    return null;
-  }
-
-  WangTile? findTileById(String id) {
-    for (final tile in tiles) {
-      if (tile.id == id) return tile;
-    }
-    return null;
-  }
-}
-
-/// Standalone tile for terrain features (forest, hills, mountain, swamp).
-class StandaloneTile {
-  final String tileId;
-  final ui.Image image;
-
-  StandaloneTile({required this.tileId, required this.image});
-}
-
-String? terrainVariantTileKey({
-  required TerrainType terrain,
-  String? resourceId,
-  int? improvementLevel,
-}) {
-  switch (terrain) {
-    case TerrainType.plains:
-      switch (resourceId) {
-        case 'grain':
-          return _tilePlainsGrain;
-        case 'meat':
-          return _tilePlainsMeat;
-        case 'horses':
-          return _tilePlainsHorses;
-        default:
-          return null;
-      }
-    case TerrainType.desert:
-      return null;
-    case TerrainType.hardwoodForest:
-      return resourceId == 'timber'
-          ? _tileHardwoodForestTimber
-          : _tileHardwoodForest;
-    case TerrainType.scrubForest:
-      return resourceId == 'timber'
-          ? _tileScrubForestTimber
-          : _tileScrubForest;
-    case TerrainType.hills:
-      if ((improvementLevel ?? 0) > 0 && _isMineResourceId(resourceId)) {
-        return _tileHillsMine;
-      }
-      return resourceId == 'wool' ? _tileHillsWool : _tileHills;
-    case TerrainType.mountain:
-      return _tileMountain;
-    case TerrainType.swamp:
-      return _tileSwamp;
-  }
-}
-
-/// L1 interior plains cells only: standalone tile key when a resource variant
-/// applies. Caller must not use this on plains↔desert transition cells (Wang).
-/// Returns null when the canonical plains Wang base should be drawn.
-String? landInteriorPlainsVariantTileKey(CellViewData cell) {
-  if (cell.terrainType != TerrainType.plains) return null;
-  return terrainVariantTileKey(
-    terrain: TerrainType.plains,
-    resourceId: cell.resourceId,
-    improvementLevel: cell.improvementLevel,
-  );
-}
-
-String featureOverlayTileKey({
-  required TerrainType terrain,
-  String? resourceId,
-  int? improvementLevel,
-}) {
-  if (terrain == TerrainType.plains || terrain == TerrainType.desert) {
-    throw UiValidationException(
-      'featureOverlayTileKey only supports L2+ feature terrains',
-    );
-  }
-  final key = terrainVariantTileKey(
-    terrain: terrain,
-    resourceId: resourceId,
-    improvementLevel: improvementLevel,
-  );
-  if (key == null) {
-    throw UiValidationException(
-      'featureOverlayTileKey only supports L2+ feature terrains',
-    );
-  }
-  return key;
-}
-
-bool _isMineResourceId(String? resourceId) {
-  switch (resourceId) {
-    case 'iron':
-    case 'copper':
-    case 'coal':
-    case 'silver':
-    case 'gold':
-    case 'gems':
-    case 'diamonds':
-    case 'tin':
-      return true;
-    default:
-      return false;
-  }
-}
 
 /// Cache for loaded tilesets.
 /// Loads L0/L1 Wang tilesets (sea_plains, sea_desert, plains_desert)
